requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
beautifulsoup4==4.12.3
certifi==2024.12.14
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
easyocr==1.7.2
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
fsspec==2024.12.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.26.1
lmdb==1.6.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
more-itertools==8.14.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip-autoremove==0.10.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
Send2Trash==1.8.3
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
-e ./__workspace_packages__
-e ./__project_packages__